Fix win32 issue

+++ tests/function-test.sh
@@ -19,9 +19,11 @@ mkdir -p ${workpath} || csih_error "Make workpath FAILED"
 csih_inform "Clean pyarmor data"
 rm -rf  ~/.pyarmor ~/.pyarmor_capsule.*
 PYARMOR_DATA=~/.pyarmor
+if [[ "${PLATFORM}" == "win_amd64" ]] ; then
 [[ -n "$USERPROFILE" ]] \
     && rm -rf "$USERPROFILE\\.pyarmor" "$USERPROFILE\\.pyarmor_capsule.*" \
     && PYARMOR_DATA=$USERPROFILE/.pyarmor
+fi
 csih_inform "PyArmor data at ${PYARMOR_DATA}"
 
 [[ -d data ]] || csih_error "No path 'data' found in current path"
